ENH: moved conversion utilities
Moved the utilities to convert from a pysat.Instrument to an xarray object to a separate sub-module.
